Move time utils to util directory
